+++ Libplanet.Tests/Common/Action/DetectRehearsal.cs
@@ -1,5 +1,5 @@
 using System.Collections.Generic;
-using System.Collections.Immutable;
+using Bencodex.Types;
 using Libplanet.Action;
 
 namespace Libplanet.Tests.Common.Action
@@ -7,27 +7,35 @@ namespace Libplanet.Tests.Common.Action
     [ActionType("detect_rehearsal")]
     public class DetectRehearsal : BaseAction
     {
-        public override IImmutableDictionary<string, object> PlainValue =>
-            new Dictionary<string, object>()
+        public override IValue PlainValue =>
+            new Bencodex.Types.Dictionary(new Dictionary<IKey, IValue>
             {
-                { "target_address", TargetAddress.ToByteArray() },
-            }.ToImmutableDictionary();
+                { (Text)"target_address", new Binary(TargetAddress.ToByteArray()) },
+            });
 
         public bool ResultState { get; set; }
 
         public Address TargetAddress { get; set; }
 
         public override void LoadPlainValue(
-            IImmutableDictionary<string, object> plainValue)
+            IValue plainValue)
         {
-            TargetAddress = new Address((byte[])plainValue["target_address"]);
+            LoadPlainValue((Bencodex.Types.Dictionary)plainValue);
+        }
+
+        public void LoadPlainValue(Bencodex.Types.Dictionary plainValue)
+        {
+            TargetAddress = new Address(plainValue.GetValue<Binary>("target_address").Value);
         }
 
         public override IAccountStateDelta Execute(IActionContext context)
         {
             IAccountStateDelta previousStates = context.PreviousStates;
             ResultState = context.Rehearsal;
-            return previousStates.SetState(TargetAddress, context.Rehearsal);
+            return previousStates.SetState(
+                TargetAddress,
+                new Bencodex.Types.Boolean(context.Rehearsal)
+            );
         }
     }
 }